* Remove 'Indicator' selector
This is temporary so I can focus on the data for now, not on the
interactivity.

I'm hard-wiring the 'emissions' dataset because it has the column
'benchmark' which is a good candidate to split the data by to
create the arrow dataset (although the name of the column in the
real data may be different).