Update video.yml
I would like to have these videos on the page "welcome" under the header "welcome to tilt's online portal. It would be great if this comes at the very end of the paragraph. 

The same holds for the second video. Would be great to have it after the body of the method section.