Welcome now includes videos (intro and method) (#193)
* Update video.yml

I would like to have these videos on the page "welcome" under the header "welcome to tilt's online portal. It would be great if this comes at the very end of the paragraph. 

The same holds for the second video. Would be great to have it after the body of the method section.

* Use "embed" URL

* Use "embed" URL

* Add videos to welcome

* Remove videos from video

* Move iframe_card() to R/iframe

* Prune needless R/video.R

* Test error

* wip

* wip

* Rename to 'Methodology'

* Use methodology text

* Rearrange cards

* refactor

* Refactor

---------

Co-authored-by: Mauro Lepore <maurolepore@gmail.com>